Disable fiber-monitoring on native
