+++ core/jvm-native/src/main/scala/cats/effect/unsafe/FiberMonitor.scala
@@ -20,8 +20,6 @@ package unsafe
 import cats.effect.tracing.TracingConstants
 import cats.effect.unsafe.ref.WeakReference
 
-import scala.concurrent.ExecutionContext
-
 import java.util.concurrent.ConcurrentLinkedQueue
 
 /**
@@ -208,15 +206,7 @@ private[effect] final class NoOpFiberMonitor extends FiberMonitor(null) {
   override def liveFiberSnapshot(print: String => Unit): Unit = {}
 }
 
-private[effect] object FiberMonitor {
-  def apply(compute: ExecutionContext): FiberMonitor = {
-    if (TracingConstants.isStackTracing && compute.isInstanceOf[WorkStealingThreadPool]) {
-      val wstp = compute.asInstanceOf[WorkStealingThreadPool]
-      new FiberMonitor(wstp)
-    } else {
-      new FiberMonitor(null)
-    }
-  }
+private[effect] object FiberMonitor extends FiberMonitorCompanionPlatform {
 
   private[FiberMonitor] final val BagReferences
       : ConcurrentLinkedQueue[WeakReference[WeakBag[Runnable]]] =
